+++ src/assets/imges/index.ts
@@ -3,4 +3,3 @@ export { default as Profilepicture } from "./profilepicture.jpeg";
 export { default as Backgroundimage } from "./backgroundimage.png";
 export { default as Skype } from "./skype.svg";
 export { default as Logo } from "./logo.svg";
-export { default as Resumes } from "./resume.png";
