+++ src/components/index.ts
@@ -1,8 +1,7 @@
 export * from "./Header";
 export * from "./Navbar";
 export * from "./About";
-export * from "./Experience";
+export * from "./Skills";
 export * from "./Portfolio";
-export * from "./Resume";
 export * from "./Footer";
-export * from "./Work";
+export * from "./Experience";
